[GREEN] TwentyPercent TwoProduct OK

+++ Logic.Interfaces/IPromotionStrategy.cs
@@ -4,5 +4,5 @@ namespace Logic.Interfaces;
 
 public interface IPromotionStrategy
 {
-    public int GetDiscount(List<Product> products);
+    public float GetDiscount(List<Product> products);
 }
